backend/requirements.txt
# AssetLens Backend Dependencies

# ============================================================================
# Web Framework
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================================================
# Database & ORM
# ============================================================================
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# ============================================================================
# Caching & Queue
# ============================================================================
redis==5.0.1
hiredis==2.3.2
celery==5.3.6

# ============================================================================
# Web Scraping
# ============================================================================
scrapy==2.11.0
scrapy-playwright==0.0.34
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
parsel==1.8.1

# ============================================================================
# HTTP Clients
# ============================================================================
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# ============================================================================
# Data Processing
# ============================================================================
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# ============================================================================
# Machine Learning
# ============================================================================
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
joblib==1.3.2

# ============================================================================
# String Matching & Deduplication
# ============================================================================
rapidfuzz==3.6.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# ============================================================================
# Date & Time Handling
# ============================================================================
python-dateutil==2.8.2
pytz==2023.3.post1
arrow==1.3.0

# ============================================================================
# Email
# ============================================================================
sendgrid==6.11.0
jinja2==3.1.3
premailer==3.10.0

# ============================================================================
# Environment & Configuration
# ============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1

# ============================================================================
# Validation & Parsing
# ============================================================================
validators==0.22.0
phonenumbers==8.13.27
ukpostcodeparser==1.1.2

# ============================================================================
# Logging & Monitoring
# ============================================================================
loguru==0.7.2
sentry-sdk==1.39.2

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0
factory-boy==3.3.0
httpx-mock==0.1.2

# ============================================================================
# Code Quality
# ============================================================================
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pylint==3.0.3

# ============================================================================
# Utilities
# ============================================================================
python-multipart==0.0.6
python-slugify==8.0.1
tenacity==8.2.3
tqdm==4.66.1
click==8.1.7
rich==13.7.0

# ============================================================================
# Security
# ============================================================================
passlib==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# ============================================================================
# API Documentation
# ============================================================================
# FastAPI includes built-in Swagger/OpenAPI support
